added config option discard-auto-indented-whitespace
Defaults to true. When disabled, the whitespace created for indentation will not be discarded if the line is left blank when exiting insert mode or pressing return. This prevents having to re-indent lines initially created as space holders before or after adding the desired text. This also helps visually to see indentation when rendering all whitespace is enabled.